requirements.txt
Django==2.1.5
django-crispy-forms==1.7.2
django-cuser==2017.3.16
html5lib==1.0.1
Pillow==5.3.0
pytz==2018.6
six==1.11.0
webencodings==0.5.1
whitenoise==4.1

